Remove excessive binaries check
* Replace kubelet cordon CLI call with client code
