+++ pkg/scripts/testdata/TestDrainNode.golden
@@ -1,5 +1,4 @@
 set -xeu pipefail
 export "PATH=$PATH:/sbin:/usr/local/bin:/opt/bin"
-
 sudo KUBECONFIG=/etc/kubernetes/admin.conf \
-    kubectl drain testNode1 --ignore-daemonsets --delete-local-data
+kubectl drain testNode1 --ignore-daemonsets --delete-local-data
