+++ pkg/scripts/testdata/TestKubeadmCentOS-v1.16.1.golden
@@ -110,9 +110,9 @@ sudo systemctl enable --now docker
 
 
 sudo yum install -y \
-	kubelet-v1.16.1 \
-	kubeadm-v1.16.1 \
-	kubectl-v1.16.1 \
+	kubelet-1.16.1 \
+	kubeadm-1.16.1 \
+	kubectl-1.16.1 \
 	kubernetes-cni-0.8.7
 sudo yum versionlock add kubelet kubeadm kubectl kubernetes-cni
