--- pkg/scripts/testdata/TestUncordonNode.golden
@@ -1,5 +1,0 @@
-set -xeu pipefail
-export "PATH=$PATH:/sbin:/usr/local/bin:/opt/bin"
-
-sudo KUBECONFIG=/etc/kubernetes/admin.conf \
-    kubectl uncordon testNode2
