--- pkg/scripts/testdata/TestVerifyPrerequisites.golden
@@ -1,15 +1,0 @@
-set -xeu pipefail
-export "PATH=$PATH:/sbin:/usr/local/bin:/opt/bin"
-
-# Check is Docker installed
-if ! type docker &>/dev/null; then exit 1; fi
-# Check is Kubelet installed
-if ! type kubelet &>/dev/null; then exit 1; fi
-# Check is Kubeadm installed
-if ! type kubeadm &>/dev/null; then exit 1; fi
-# Check do Kubernetes directories and files exist
-if [[ ! -d "/etc/kubernetes/manifests" ]]; then exit 1; fi
-if [[ ! -d "/etc/kubernetes/pki" ]]; then exit 1; fi
-if [[ ! -f "/etc/kubernetes/kubelet.conf" ]]; then exit 1; fi
-# Check are kubelet running
-if ! sudo systemctl is-active --quiet kubelet &>/dev/null; then exit 1; fi
